% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_function_verse.R
\name{annotatePathways}
\alias{annotatePathways}
\title{Annotate pathways for input data}
\usage{
annotatePathways(species, selected.db, input.data)
}
\arguments{
\item{species}{}

\item{selected.db}{}

\item{input.data}{}
}
\value{

}
\description{
Annotate pathways for input data
}
