% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_upload.R
\name{checkLL_RR}
\alias{checkLL_RR}
\title{Manually change the annotation of L-L and R-R pairs}
\usage{
checkLL_RR(input.data)
}
\arguments{
\item{input.data}{preprocessed table}
}
\value{
input.data
}
\description{
Manually change the annotation of L-L and R-R pairs
}
\examples{
data(input.data)
checked.input.data <- checkLL_RR(input.data)

}
