% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_int_pair_modules.R
\name{circlePlot}
\alias{circlePlot}
\title{Plot circle plot}
\usage{
circlePlot(data, cluster_colors, int_flow, link.color)
}
\arguments{
\item{data}{}

\item{cluster_colors}{}

\item{int_flow}{}

\item{link.color}{}
}
\value{

}
\description{
Plot circle plot
}
