% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_function_verse.R
\name{combineAnnotations}
\alias{combineAnnotations}
\title{Combine GO annotation and pathways in a unique object}
\usage{
combineAnnotations(GO_annotation, pathways_annotation)
}
\arguments{
\item{GO_annotation}{data}

\item{pathways_annotation}{data}
}
\value{
combined annotation dataframe
}
\description{
Combine GO annotation and pathways in a unique object
}
