% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_verse.R
\name{createBarPlot2_CV}
\alias{createBarPlot2_CV}
\title{Create barplot of number of interaction for selected cluster}
\usage{
createBarPlot2_CV(
  filt.data,
  input_cluster_selected_checkbox,
  input_clust_barplot2
)
}
\arguments{
\item{filt.data}{input data filtered in cluster-verse}

\item{input_cluster_selected_checkbox}{selected clusters to keep}

\item{input_clust_barplot2}{selected cluster to plot}
}
\value{
plotly fig
}
\description{
Create barplot of number of interaction for selected cluster
}
