% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_verse.R
\name{createBarPlot2_ggplot}
\alias{createBarPlot2_ggplot}
\title{Create ggplot barplot of Nint per cluster selected}
\usage{
createBarPlot2_ggplot(
  barplotDF2,
  input_cluster_selected_checkbox,
  input_clust_barplot2
)
}
\arguments{
\item{barplotDF2}{dataframe with barplot data}

\item{input_cluster_selected_checkbox}{selected clusters to keep}

\item{input_clust_barplot2}{selected cluster to plot}
}
\value{
ggplot barplot
}
\description{
Create ggplot barplot of Nint per cluster selected
}
