% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_verse.R
\name{createBarPlot_CV}
\alias{createBarPlot_CV}
\title{Create Barplot cluster-verse}
\usage{
createBarPlot_CV(
  barplotDF,
  input_cluster_selected_checkbox,
  input_num_or_weight_bar1
)
}
\arguments{
\item{barplotDF}{dataframe with N interactions per cluster (auto/para)}

\item{input_cluster_selected_checkbox}{checkbox input}

\item{input_num_or_weight_bar1}{number of int or weighted number by score}
}
\value{
plotly barplot
}
\description{
Create Barplot cluster-verse
}
