% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_int_pair_modules.R
\name{elbowPoint}
\alias{elbowPoint}
\title{Determine the elbow point on a curve (from package akmedoids)}
\usage{
elbowPoint(x, y)
}
\arguments{
\item{x}{vector of x coordinates of points on the curve}

\item{y}{vector of y coordinates of points on the curve}
}
\value{
an x, y coordinates of the elbow point.
}
\description{
Given a list of x, y coordinates on a curve, function determines the elbow point of the curve.
}
\details{
highlight the maximum curvature to identify the elbow point (credit: 'github.com/agentlans')
}
