% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gene_verse.R
\name{ensemblLink}
\alias{ensemblLink}
\title{Get html link to ensembl}
\usage{
ensemblLink(ensembl)
}
\arguments{
\item{ensembl}{symbol}
}
\value{
html link to website
}
\description{
Get html link to ensembl
}
