% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multi_cond.R
\name{getBack2BackBarplot}
\alias{getBack2BackBarplot}
\title{Get back-to-back barplot for 2 conditions comparison}
\usage{
getBack2BackBarplot(tab_c1, tab_c2, lab_c1, lab_c2)
}
\arguments{
\item{tab_c1}{barplot dataframe generated by getBarplotDF() for condition 1}

\item{tab_c2}{barplot dataframe generated by getBarplotDF() for condition 1}

\item{lab_c1}{label for condition 1}

\item{lab_c2}{label for condition 2}
}
\value{
ggplot object
}
\description{
Get back-to-back barplot for 2 conditions comparison
}
