% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_verse.R
\name{getBarplotDF}
\alias{getBarplotDF}
\title{Get dataframe for plotting barplot (all clusters)}
\usage{
getBarplotDF(
  data.filt.bar,
  input_cluster_selected_checkbox,
  input_num_or_weight_bar1
)
}
\arguments{
\item{data.filt.bar}{filtered object (checkbox auto/para)}

\item{input_cluster_selected_checkbox}{checkbox input}

\item{input_num_or_weight_bar1}{number of int or weighted number by score}
}
\value{
dataframe with number of interactions per cluster auto/para
}
\description{
Get dataframe for plotting barplot (all clusters)
}
