% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_verse.R
\name{getBarplotDF2}
\alias{getBarplotDF2}
\title{Get dataframe for barplot (by cluster)}
\usage{
getBarplotDF2(filt.data, input_cluster_selected_checkbox, input_clust_barplot2)
}
\arguments{
\item{filt.data}{input data filtered in cluster-verse}

\item{input_cluster_selected_checkbox}{selected clusters to keep}

\item{input_clust_barplot2}{selected cluster to plot}
}
\value{
dataframe with num int per cluster
}
\description{
Get dataframe for barplot (by cluster)
}
