% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_verse.R
\name{getClusterA_Names}
\alias{getClusterA_Names}
\title{Get cluster names only from sender cluster A}
\usage{
getClusterA_Names(input.data)
}
\arguments{
\item{input.data}{preprocessed input data}
}
\value{
named list of clusters
}
\description{
Get cluster names only from sender cluster A
}
