% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_verse.R
\name{getClusterColors}
\alias{getClusterColors}
\title{Get colors for clusters}
\usage{
getClusterColors(input.data)
}
\arguments{
\item{input.data}{preprocessed input data}
}
\value{
named vector with colors per cluster
}
\description{
Get colors for clusters
}
