% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_verse.R
\name{getClusterNetwork}
\alias{getClusterNetwork}
\title{Creating edges dataframe for network of clusters}
\usage{
getClusterNetwork(input.data, input_num_or_weight_radio, input_edge_weight)
}
\arguments{
\item{input.data}{preprocessed input data}

\item{input_num_or_weight_radio}{either num of interactions or weighted by score}

\item{input_edge_weight}{small,medium or large from user input}
}
\value{
edges dataframe
}
\description{
Creating edges dataframe for network of clusters
}
