% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_verse.R
\name{getClusterSize}
\alias{getClusterSize}
\title{Get Clusters size}
\usage{
getClusterSize(cl, edges.df)
}
\arguments{
\item{cl}{cluster name}

\item{edges.df}{dataframe with edges for network}
}
\value{
sum of interactions for that cluster
}
\description{
Get Clusters size
}
