% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multi_cond.R
\name{getDistinctCouplets}
\alias{getDistinctCouplets}
\title{Get table of unique int-pairs/clust-pairs couplets}
\usage{
getDistinctCouplets(
  data_cond1,
  data_cond2,
  data_cond3 = NULL,
  lab_c1,
  lab_c2,
  lab_c3 = NULL
)
}
\arguments{
\item{data_cond1}{filt.data() corresponding to chosen condition 1}

\item{data_cond2}{filt.data() corresponding to chosen condition 2}

\item{data_cond3}{filt.data() corresponding to chosen condition 3}

\item{lab_c1}{data label for condition 1}

\item{lab_c2}{data label for condition 2}

\item{lab_c3}{data label for condition 3}
}
\value{
modified filt.data containing only unique couplets
}
\description{
Get table of unique int-pairs/clust-pairs couplets
}
