% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gene_verse.R
\name{getDotPlot_selInt}
\alias{getDotPlot_selInt}
\title{Functions to plot DotPlots}
\usage{
getDotPlot_selInt(
  selected_tab,
  clust.order,
  low_color = "aquamarine",
  high_color = "#131780"
)
}
\arguments{
\item{selected_tab}{}

\item{clust.order}{}

\item{low_color}{}

\item{high_color}{}
}
\value{

}
\description{
Functions to plot DotPlots
}
