% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gene_verse.R
\name{getGeneTable}
\alias{getGeneTable}
\title{Get table for gene-verse}
\usage{
getGeneTable(input.data)
}
\arguments{
\item{input.data}{preprocessed input data}
}
\value{
gene table with unique intpairs (no connection to clusters)
}
\description{
Get table for gene-verse
}
\examples{
gene_table <- getGeneTable(input_data)
}
