% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_annotation.R
\name{getGObiomaRt}
\alias{getGObiomaRt}
\title{Connection to Ensembl via biomaRt to get GO terms}
\usage{
getGObiomaRt(input_select_ensembl, input.data)
}
\arguments{
\item{input_select_ensembl}{chosen version of Ensembl}

\item{input.data}{filtered input data}
}
\value{
dataframe with GO annotation
}
\description{
Connection to Ensembl via biomaRt to get GO terms
}
