% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_annotation.R
\name{getGraphiteDB}
\alias{getGraphiteDB}
\title{Get pathway databases from graphite package}
\usage{
getGraphiteDB(species, database)
}
\arguments{
\item{species}{for now h_sapiens}

\item{database}{name of the database to download}
}
\value{
dataframe containing the downloaded database to pre-process
}
\description{
Get pathway databases from graphite package
}
