% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_int_pair_modules.R
\name{getHitsf}
\alias{getHitsf}
\title{Subfunction to calculate significant functions by permutation test}
\usage{
getHitsf(mat, gpModules_assign)
}
\arguments{
\item{mat}{binary matrix of functional terms by int-pairs}

\item{gpModules_assign}{assignment of intpairs to modules}
}
\value{
matrix with hits
}
\description{
Subfunction to calculate significant functions by permutation test
}
