% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_verse.R
\name{getIntFlow}
\alias{getIntFlow}
\title{Get subset of interactions corresponding to a certain viewpoint and flow}
\usage{
getIntFlow(vp, input.data, flow)
}
\arguments{
\item{vp}{}

\item{input.data}{}

\item{flow}{}
}
\value{

}
\description{
Get subset of interactions corresponding to a certain viewpoint and flow
}
