% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_function_verse.R
\name{getNtermsBYdb}
\alias{getNtermsBYdb}
\title{Calculate number of terms of a database}
\usage{
getNtermsBYdb(annotation)
}
\arguments{
\item{annotation}{}
}
\value{

}
\description{
Calculate number of terms of a database
}
