% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gene_verse.R
\name{getNumLR}
\alias{getNumLR}
\title{Get number of unique ligands and receptors}
\usage{
getNumLR(input.data, type)
}
\arguments{
\item{input.data}{}

\item{type}{}
}
\value{

}
\description{
Get number of unique ligands and receptors
}
