% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_int_pair_modules.R
\name{getOccurrenceTab4wordcloud}
\alias{getOccurrenceTab4wordcloud}
\title{Construct table with occurrence values for significant terms}
\usage{
getOccurrenceTab4wordcloud(
  signFun,
  ipMselected,
  gpModules_assign,
  subGenePairs_func_mat
)
}
\arguments{
\item{signFun}{table of significant terms}

\item{ipMselected}{number of intpair module selected by user}

\item{gpModules_assign}{assignment of intpairs to modules}

\item{subGenePairs_func_mat}{subset of binary matrix}
}
\value{
table of occurrence
}
\description{
Construct table with occurrence values for significant terms
}
