% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multi_cond.R
\name{getPieChart}
\alias{getPieChart}
\title{Get Pie Chart of unique couplets}
\usage{
getPieChart(data_dotplot)
}
\arguments{
\item{data_dotplot}{same data used to generate dotplot}
}
\value{
pie chart
}
\description{
Get Pie Chart of unique couplets
}
