% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multi_cond.R
\name{getRadar_df}
\alias{getRadar_df}
\title{Get radar df of relative numbers of interactions for a certain cell type}
\usage{
getRadar_df(tab_c1, tab_c2, tab_c3, lab_c1, lab_c2, lab_c3)
}
\arguments{
\item{tab_c1}{barplot dataframe from Viewpoint generated by getBarplotDF2() containing data for condition 1}

\item{tab_c2}{barplot dataframe from Viewpoint generated by getBarplotDF2() containing data for condition 2}

\item{tab_c3}{barplot dataframe from Viewpoint generated by getBarplotDF2() containing data for condition 3}

\item{lab_c1}{label for condition 1}

\item{lab_c2}{label for condition 2}

\item{lab_c3}{label for condition 3}
}
\value{
df to be then used with fmsb radarchart
}
\description{
Get radar df of relative numbers of interactions for a certain cell type
}
