% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_int_pair_modules.R
\name{getSignificantFunctions}
\alias{getSignificantFunctions}
\title{Title}
\usage{
getSignificantFunctions(
  subGenePairs_func_mat,
  gpModules_assign,
  rank.terms,
  input_maxPval
)
}
\arguments{
\item{subGenePairs_func_mat}{}

\item{gpModules_assign}{}
}
\value{

}
\description{
Title
}
