% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multi_cond.R
\name{getSignificantFunctions_multiCond}
\alias{getSignificantFunctions_multiCond}
\title{Get significance of functional terms related to unique int-pairs per condition}
\usage{
getSignificantFunctions_multiCond(sub_annot, unique_intpairs)
}
\arguments{
\item{sub_annot}{annotation matrix subset to unique int-pairs}

\item{unique_intpairs}{data.frame with unique int-pairs by condition}
}
\value{
data.frame with calculated pvalue of significance
}
\description{
Get significance of functional terms related to unique int-pairs per condition
}
