% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_function_verse.R
\name{getSunburst}
\alias{getSunburst}
\title{Get Sunburst plot of selected functional terms}
\usage{
getSunburst(
  sel.data,
  func_selected,
  int_p_fun,
  cluster.colors,
  input_num_or_weight_radio
)
}
\arguments{
\item{sel.data}{dataframe of selected functions}

\item{func_selected}{the selected functional term}

\item{int_p_fun}{dataframe with int pairs annotated to this function}

\item{cluster.colors}{for plotting}

\item{input_num_or_weight_radio}{either num of interactions or weighted by score}
}
\value{
plotly figure
}
\description{
Get Sunburst plot of selected functional terms
}
