% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_int_pair_modules.R
\name{plotWordCloud}
\alias{plotWordCloud}
\title{Title}
\usage{
plotWordCloud(occurTab, input_minFreq_wordcloud)
}
\arguments{
\item{occurTab}{}

\item{input_minFreq_wordcloud}{}
}
\value{

}
\description{
Title
}
