% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_upload.R
\name{read.cellchat}
\alias{read.cellchat}
\title{Read dataframe of cell-cell communication from CellChat (ligand/receptor)}
\usage{
read.cellchat(file_tab)
}
\arguments{
\item{file_tab}{dataframe from cellchat}
}
\value{
input.data formatted for InterCellar
}
\description{
Read dataframe of cell-cell communication from CellChat (ligand/receptor)
}
