% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_upload_custom.R
\name{read.customInput}
\alias{read.customInput}
\title{Read custom input file and re-structure it with InterCellar format}
\usage{
read.customInput(tab, separator)
}
\arguments{
\item{tab}{}

\item{separator}{}
}
\value{

}
\description{
Read custom input file and re-structure it with InterCellar format
}
