% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_upload.R
\name{read.icellnet}
\alias{read.icellnet}
\title{Read ICELLNET dataframe}
\usage{
read.icellnet(tab, input_icellnet_CC, input_icellnet_dir)
}
\arguments{
\item{tab}{dataframe with int-pairs in "X" column, other columns as cell types}

\item{input_icellnet_CC}{central cell name}

\item{input_icellnet_dir}{direction of interaction either out or in}
}
\value{
pre-processed input data
}
\description{
Read ICELLNET dataframe
}
