% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(reproducible = TRUE)
}
\arguments{
\item{reproducible}{boolean for setting a seed, making plots reproducible}
}
\value{
a running instance of InterCellar
}
\description{
Run the Shiny Application
}
\examples{
\dontrun{
run_app()
}

}
