% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multi_cond.R
\name{subsetAnnot_multiCond}
\alias{subsetAnnot_multiCond}
\title{Subset int-pair by function matrices to unique int-pairs by condition}
\usage{
subsetAnnot_multiCond(
  annot_cond1,
  annot_cond2,
  annot_cond3,
  unique_intpairs,
  lab_c1,
  lab_c2,
  lab_c3
)
}
\arguments{
\item{annot_cond1}{binary matrix int-pair by functions for cond1}

\item{annot_cond2}{binary matrix int-pair by functions for cond2}

\item{annot_cond3}{binary matrix int-pair by functions for cond3}

\item{unique_intpairs}{table of unique int-pairs by condition}

\item{lab_c1}{label cond1}

\item{lab_c2}{label cond2}

\item{lab_c3}{label cond3}
}
\value{
subset merged matrix
}
\description{
Subset int-pair by function matrices to unique int-pairs by condition
}
