% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_int_pair_modules.R
\name{subsetFuncMatBYFlow}
\alias{subsetFuncMatBYFlow}
\title{Subset pairs-function matrix by selected flow}
\usage{
subsetFuncMatBYFlow(pairs_func_matrix, flow_df)
}
\arguments{
\item{pairs_func_matrix}{binary}

\item{flow_df}{subset of input data by flow}
}
\value{
subset of binary mat
}
\description{
Subset pairs-function matrix by selected flow
}
