% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_upload.R
\name{swap.RLint}
\alias{swap.RLint}
\title{Swaps interaction pairs that are R-L to L-R}
\usage{
swap.RLint(RLint)
}
\arguments{
\item{RLint}{subset of R-L interactions}
}
\value{
input data with ordered L-R pairs and L-L/R-R
}
\description{
Swaps interaction pairs that are R-L to L-R
}
