% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gene_verse.R
\name{uniprotLink}
\alias{uniprotLink}
\title{Get html link to uniprot}
\usage{
uniprotLink(uniprot)
}
\arguments{
\item{uniprot}{symbol}
}
\value{
html link to website
}
\description{
Get html link to uniprot
}
