% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_upload.R
\name{updateInputLR}
\alias{updateInputLR}
\title{Function that orders all interaction pairs as L-R. Leaves unchanged the
R-R and L-L}
\usage{
updateInputLR(input.data)
}
\arguments{
\item{input.data}{uploaded data}
}
\value{
ordered input data
}
\description{
Function that orders all interaction pairs as L-R. Leaves unchanged the
R-R and L-L
}
