% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_int_pair_modules.R
\name{wordcloud2FixHTML}
\alias{wordcloud2FixHTML}
\title{Title}
\usage{
wordcloud2FixHTML(inputFile, outputFile)
}
\arguments{
\item{inputFile}{}

\item{outputFile}{}
}
\value{

}
\description{
Title
}
