% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_function_verse.R
\name{annotatePathways}
\alias{annotatePathways}
\title{Annotate pathways for input data}
\usage{
annotatePathways(selected.db, input.data)
}
\arguments{
\item{selected.db}{pathways sources to use}

\item{input.data}{filtered input data}
}
\value{
pathways_annotation
}
\description{
Annotate pathways for input data
}
