% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_function_verse.R
\name{buildPairsbyFunctionMatrix}
\alias{buildPairsbyFunctionMatrix}
\title{Build binary matrix with int-pairs in rows, functions in cols}
\usage{
buildPairsbyFunctionMatrix(functions_df)
}
\arguments{
\item{functions_df}{}
}
\value{

}
\description{
Build binary matrix with int-pairs in rows, functions in cols
}
