% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_int_pair_modules.R
\name{circlePlot}
\alias{circlePlot}
\title{Plot circle plot}
\usage{
circlePlot(data, cluster_colors, ipm_color, int_flow, link.color)
}
\arguments{
\item{data}{subset of input data by flow / intpair module}

\item{cluster_colors}{global}

\item{ipm_color}{single color for chosen int-pair module}

\item{int_flow}{string specifying the flow}

\item{link.color}{string specifying variable by which to color links}
}
\value{
circle plot
}
\description{
Plot circle plot
}
