% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_verse.R
\name{createBarPlot1_ggplot}
\alias{createBarPlot1_ggplot}
\title{Create ggplot barplot to be saved in tiff}
\usage{
createBarPlot1_ggplot(
  barplotDF,
  input_cluster_selected_checkbox,
  input_num_or_weight_bar1
)
}
\arguments{
\item{barplotDF}{dataframe with N interactions per cluster (auto/para)}

\item{input_cluster_selected_checkbox}{checkbox input}

\item{input_num_or_weight_bar1}{number of int or weighted number by score}
}
\value{
ggplot barplot
}
\description{
Create ggplot barplot to be saved in tiff
}
