% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_verse.R
\name{createBarPlot_CV}
\alias{createBarPlot_CV}
\title{Create Barplot cluster-verse}
\usage{
createBarPlot_CV(data.filt.bar, input_cluster_selected_checkbox)
}
\arguments{
\item{data.filt.bar}{filtered object (checkbox auto/para)}

\item{input_cluster_selected_checkbox}{checkbox input}
}
\value{
plotly barplot
}
\description{
Create Barplot cluster-verse
}
