% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_verse.R
\name{createNetwork}
\alias{createNetwork}
\title{Create Network of clusters}
\usage{
createNetwork(data.filt.cluster, input_num_or_weight_radio)
}
\arguments{
\item{data.filt.cluster}{filtered input data (by clusters)}

\item{input_num_or_weight_radio}{either number of interactions or weighted by score}
}
\value{
list containing nodes and edges for network
}
\description{
Create Network of clusters
}
