% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_verse.R
\name{createNetwork}
\alias{createNetwork}
\title{Create Network of clusters}
\usage{
createNetwork(data.filt.cluster)
}
\arguments{
\item{data.filt.cluster}{}
}
\value{

}
\description{
Create Network of clusters
}
