% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_int_pair_modules.R
\name{dendroIntPairModules}
\alias{dendroIntPairModules}
\title{Get dendrogram of int pair modules}
\usage{
dendroIntPairModules(pairs_func_matrix, seed = NULL)
}
\arguments{
\item{pairs_func_matrix}{}

\item{seed}{}
}
\value{

}
\description{
Get dendrogram of int pair modules
}
