% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_verse.R
\name{getClusterNetwork}
\alias{getClusterNetwork}
\title{Creating edges dataframe for network of clusters}
\usage{
getClusterNetwork(input.data)
}
\arguments{
\item{input.data}{preprocessed input data}
}
\value{
edges dataframe
}
\description{
Creating edges dataframe for network of clusters
}
