% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_verse.R
\name{getClusterNetwork}
\alias{getClusterNetwork}
\title{Creating edges dataframe for network of clusters}
\usage{
getClusterNetwork(input.data, input_num_or_weight_radio)
}
\arguments{
\item{input.data}{preprocessed input data}

\item{input_num_or_weight_radio}{either num of interactions or weighted by score}
}
\value{
edges dataframe
}
\description{
Creating edges dataframe for network of clusters
}
