% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gene_verse.R
\name{getDotPlot_selInt}
\alias{getDotPlot_selInt}
\title{Functions to plot DotPlots}
\usage{
getDotPlot_selInt(
  selected_tab,
  clust.order,
  low_color = "aquamarine",
  high_color = "#131780"
)
}
\arguments{
\item{selected_tab}{table of selected rows from gene tableeeeeee}

\item{clust.order}{how to order clusters}

\item{low_color}{of dotplot}

\item{high_color}{of dotplot}
}
\value{
list with modified selected data and ggplot2 dotplot
}
\description{
Functions to plot DotPlots
}
