% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_verse.R
\name{getIntFlow}
\alias{getIntFlow}
\title{Get subset of interactions corresponding to a certain viewpoint and flow}
\usage{
getIntFlow(vp, input.data, flow)
}
\arguments{
\item{vp}{viewpoint cluster}

\item{input.data}{preprocessed/filtered input data}

\item{flow}{one among directed_out, directed_in or undirected}
}
\value{
subset of data
}
\description{
Get subset of interactions corresponding to a certain viewpoint and flow
}
\examples{
data(input.data)
caf_out <- getIntFlow(vp = "CAF", input.data, flow = "directed_out")
}
