% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_function_verse.R
\name{getNtermsBYdb}
\alias{getNtermsBYdb}
\title{Calculate number of terms of a database}
\usage{
getNtermsBYdb(annotation)
}
\arguments{
\item{annotation}{data from either pathways, GO or combined}
}
\value{
number of terms by dataset
}
\description{
Calculate number of terms of a database
}
