% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gene_verse.R
\name{getNumLR}
\alias{getNumLR}
\title{Get number of unique ligands and receptors}
\usage{
getNumLR(input.data, type)
}
\arguments{
\item{input.data}{preprocessed input data}

\item{type}{either L or R}
}
\value{
number of L or R genes
}
\description{
Get number of unique ligands and receptors
}
