% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_int_pair_modules.R
\name{getOccurrenceTab4wordcloud}
\alias{getOccurrenceTab4wordcloud}
\title{Title}
\usage{
getOccurrenceTab4wordcloud(
  signFun,
  ipMselected,
  gpModules_assign,
  subGenePairs_func_mat
)
}
\arguments{
\item{signFun}{}

\item{ipMselected}{}

\item{gpModules_assign}{}

\item{subGenePairs_func_mat}{}
}
\value{

}
\description{
Title
}
