% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cluster_verse.R
\name{getRadarPlot}
\alias{getRadarPlot}
\title{Get radar plot of relative numbers of interactions for a certain cell type}
\usage{
getRadarPlot(tab_c1, tab_c2, lab_c1, lab_c2, cell_name)
}
\arguments{
\item{tab_c1}{table from csv file (barplot#2) containing data for condition 1}

\item{tab_c2}{table from csv file (barplot#2) containing data for condition 2}

\item{lab_c1}{label for condition 1}

\item{lab_c2}{label for condition 2}

\item{cell_name}{label of cell type of interest}
}
\value{
plot
}
\description{
Get radar plot of relative numbers of interactions for a certain cell type
}
