% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_function_verse.R
\name{getRankedTerms}
\alias{getRankedTerms}
\title{Get table with ranked functional terms}
\usage{
getRankedTerms(data.fun.annot, gene.table)
}
\arguments{
\item{data.fun.annot}{}
}
\value{

}
\description{
Get table with ranked functional terms
}
