% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_function_verse.R
\name{getRankedTerms}
\alias{getRankedTerms}
\title{Get table with ranked functional terms}
\usage{
getRankedTerms(data.fun.annot)
}
\arguments{
\item{data.fun.annot}{annotated df (GO/path/combined)}
}
\value{
table with ranking
}
\description{
Get table with ranked functional terms
}
