% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_int_pair_modules.R
\name{getUMAPipModules}
\alias{getUMAPipModules}
\title{Get UMAP for IP modules}
\usage{
getUMAPipModules(intPairs.dendro, gpModules_assign, ipm_colors)
}
\arguments{
\item{intPairs.dendro}{list output of dendrogram}

\item{gpModules_assign}{named vector of module assignment}

\item{ipm_colors}{for intpair modules}
}
\value{
plotly umap
}
\description{
Get UMAP for IP modules
}
