% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_int_pair_modules.R
\name{getUMAPipModules}
\alias{getUMAPipModules}
\title{Get UMAP for IP modules}
\usage{
getUMAPipModules(
  intPairs.dendro,
  gpModules_assign,
  gene.table,
  ipm_colors,
  input_ipM_UMAPcolors
)
}
\arguments{
\item{intPairs.dendro}{}

\item{gpModules_assign}{}

\item{ipm_colors}{}

\item{input_ipM_UMAPcolors}{}
}
\value{

}
\description{
Get UMAP for IP modules
}
