% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_gene_verse.R
\name{getUniqueDotplot}
\alias{getUniqueDotplot}
\title{Plot dotplot containing only unique int-pair/cluster pairs with many conditions}
\usage{
getUniqueDotplot(data_dotplot)
}
\arguments{
\item{data_dotplot}{table with selected int_pairs for multiple conditions}
}
\value{
ggplot object
}
\description{
Plot dotplot containing only unique int-pair/cluster pairs with many conditions
}
