% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_int_pair_modules.R
\name{plotWordCloud}
\alias{plotWordCloud}
\title{Plot wordcloud of significant terms}
\usage{
plotWordCloud(sign_table, input_minFreq_wordcloud)
}
\arguments{
\item{sign_table}{table with significant terms for ipm chosen}

\item{input_minFreq_wordcloud}{threshold on min occurrence by user}
}
\value{
wordloud2 html object
}
\description{
Plot wordcloud of significant terms
}
