% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_int_pair_modules.R
\name{wordcloud2FixHTML}
\alias{wordcloud2FixHTML}
\title{Function to fix the html generated by wordcloud2}
\usage{
wordcloud2FixHTML(inputFile, outputFile)
}
\arguments{
\item{inputFile}{html file}

\item{outputFile}{html file}
}
\value{
fixed html
}
\description{
Function to fix the html generated by wordcloud2
}
